Don't explode when a rel attribute exists without a base tag.
It tries to build an absolute URL for link tags that contain a rel
attribute and a relative URL, using the <base> tag to decide what to
prefix the relative URLs with.  It's very common for people to omit the
base tag and to use rel attributes in stylesheets, and that blows up the
parser.

Now we don't attempt to build absolute URLs when the base URL can't be
determined.